# specs/001-modulo-stock-pedidos/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Stock.Api — Módulo de Stock y generación automática de pedidos
  version: 1.0.0
  description: |
    API REST del módulo de stock. Ver `README.md` para convenciones de autenticación,
    autorización y mapeo de códigos de estado.
    Toda regla de negocio vive detrás de esta API; `Stock.Web` es sólo un consumidor.
servers:
  - url: http://localhost:5279
    description: Desarrollo local vía docker compose
security:
  - bearerAuth: []
...
paths:
...
  /api/articulos:
    get:
      tags: [Articulos]
      summary: Listar artículos, con filtro opcional por descripción
      description: >
        Devuelve como máximo 10.000 filas ordenadas por Código ascendente. El tope no es opcional:
        la constitución prohíbe toda consulta sin límite ni filtro. Es el endpoint que consume la
        ventana de búsqueda de artículos. (RF-027, RF-034a)
      parameters:
        - name: descripcion
          in: query
          schema: { type: string }
          description: Coincidencia parcial, insensible a mayúsculas y acentos. (RF-027a)
        - name: codigo
          in: query
          schema: { type: string }
          description: >
            Coincidencia **exacta**, con la misma regla de comparación que la unicidad del Código:
            insensible a mayúsculas y sensible a acentos (RF-017a). Devuelve un arreglo de 0 ó 1
            elementos; un Código inexistente responde **200 con arreglo vacío y no 404**, porque
            para la pantalla de carga eso significa "no hay sugerencia" y no un error (RF-020g). El
            404 por Código inexistente es el de RF-020e, al grabar el movimiento.

            Es la resolución por Código que consume la pantalla de movimientos —una sola consulta
            para la Descripción y los dos precios— y se combina con `descripcion` sin reemplazarla:
            los dos filtros se acumulan.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Articulo' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
    post:
      tags: [Articulos]
      summary: Alta de artículo (RF-013)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ArticuloRequest' }
      responses:
        '201':
          description: Creado. `precioVenta` viene calculado por el sistema. (RF-016)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Articulo' }
        '400':
          description: |
            Valor negativo, parámetro de reposición no entero, o incumplimiento de
            StockMinimo ≤ PuntoPedido ≤ StockIdeal. (RF-018, RF-019)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '409':
          description: Código de artículo duplicado. (RF-017)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
...
  /api/movimientos:
    get:
      tags: [Movimientos]
      summary: Listar movimientos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Movimiento' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
    post:
      tags: [Movimientos]
      summary: Alta de movimiento (RF-020). Todo-o-nada. (RF-024c)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MovimientoRequest' }
      responses:
        '201':
          description: |
            Creado. `numero` lo asigna el sistema desde una secuencia global única
            compartida entre compras y ventas. (RF-020a)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Movimiento' }
        '400':
          description: |
            Cantidad no entera (rechazada al deserializar, RF-018a); cantidad ≤ 0
            o > 1.000.000; Precio Unitario negativo o fuera de tope; fecha futura;
            tipo inválido. (RF-018a, RF-020b, RF-020d, RF-023, RF-023a, RF-023c)
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/ProblemDetails' }
        '401': { $ref: '#/components/responses/NoAutorizado' }
        '422': { $ref: '#/components/responses/StockInsuficiente' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NoAutorizado:
      description: Sin sesión autenticada válida. (RF-012)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    StockInsuficiente:
      description: |
        La operación dejaría el Stock Actual de algún artículo por debajo de 0.
        Es también la respuesta ante concurrencia perdida, evaluada contra el saldo
        ya actualizado: nunca se devuelve un error de reintento. (RF-024, RF-024b)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
  schemas:
    ProblemDetails:
      type: object
      description: RFC 7807. `detail` es apto para mostrar al usuario; nunca expone detalle interno.
      properties:
        type:   { type: string }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string, example: 'Stock insuficiente para el artículo A-001.' }
...
    Articulo:
      type: object
      properties:
        articuloId:   { type: integer }
        codigo:       { type: string }
        descripcion:  { type: string }
        precioCosto:  { type: number, format: double }
        margen:       { type: number, format: double, description: Porcentaje. }
        precioVenta:  { type: number, format: double, readOnly: true, description: 'Calculado: costo × (1 + margen/100). (RF-016)' }
        stockMinimo:  { type: integer }
        puntoPedido:  { type: integer }
        stockIdeal:   { type: integer }
    ArticuloRequest:
      type: object
      required: [codigo, descripcion, precioCosto, margen, stockMinimo, puntoPedido, stockIdeal]
      properties:
        codigo:       { type: string, maxLength: 50 }
        descripcion:  { type: string, maxLength: 200 }
        precioCosto:  { type: number, format: double, minimum: 0 }
        margen:       { type: number, format: double, minimum: 0 }
        stockMinimo:  { type: integer, minimum: 0, description: Entero no negativo. (RF-013a) }
        puntoPedido:  { type: integer, minimum: 0 }
        stockIdeal:   { type: integer, minimum: 0 }
    TipoMovimiento:
      type: string
      enum: [Compra, Venta]
      description: Conjunto cerrado. Compra suma al stock, Venta resta. (RF-020b)
    Movimiento:
      type: object
      properties:
        numero: { type: integer, readOnly: true, description: Asignado por el sistema. (RF-020a) }
        tipo:   { $ref: '#/components/schemas/TipoMovimiento' }
        fecha:  { type: string, format: date }
        detalle:
          type: array
          items: { $ref: '#/components/schemas/MovimientoDetalle' }
    MovimientoDetalle:
      type: object
      properties:
        # El Código es la identidad de negocio de la línea, en la respuesta y en el pedido
        # (RF-020, RF-020e). El identificador interno del artículo no cruza el borde de la API:
        # es la referencia física del modelo de datos y exponerlo obligaría al cliente a
        # conocerlo para operar.
        codigo:         { type: string }
        cantidad:       { type: integer, minimum: 1, maximum: 1000000 }
        precioUnitario: { type: number, format: double, minimum: 0 }
        precioTotal:    { type: number, format: double, readOnly: true, description: 'Calculado: cantidad × precio unitario. (RF-020c)' }
    MovimientoRequest:
      type: object
      required: [tipo, fecha, detalle]
      properties:
        tipo:   { $ref: '#/components/schemas/TipoMovimiento' }
        fecha:  { type: string, format: date, description: No puede ser posterior a hoy. (RF-020d) }
        detalle:
          type: array
          minItems: 1
          items:
            type: object
            required: [codigo, cantidad, precioUnitario]
            properties:
              # Código y no identificador interno (RF-020e). Un Código inexistente se responde
              # 404, con la misma regla de comparación insensible a mayúsculas de RF-017a.
              codigo:         { type: string, maxLength: 50 }
              cantidad:       { type: integer, minimum: 1, maximum: 1000000 }
              precioUnitario: { type: number, format: double, minimum: 0 }
...